Refactoring
* Creating character class for CTraveller creation
* Adding it to cli
* Cleanup of idea config files
